rename home to mylocation
