--- app/src/main/java/com/example/weatherapp/feature/home/util/HomePageState.kt
@@ -1,11 +1,0 @@
-package com.example.weatherapp.feature.home.util
-
-import com.example.weatherapp.feature.favouritelocations.util.Coordinates
-
-/**
- * State holder for Home page screen.
- */
-class HomePageState {
-    var isLocationAlreadyUpdated = false
-    var retrievedCoordinates: Coordinates? = null
-}
